requirements.txt
aiogram
python-dotenv
requests
sqlalchemy
aiohttp
SQLAlchemy